с fix: fixed header, split header and home, add router
