+++ src/App.css
@@ -3,7 +3,7 @@
 }
 
 .Header-row {
-  align-items: center;
+  align-items: middle;
 }
 
 .Header-pane {
